# internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rebecca API",
    "version": "0.2.0",
    "description": "Go-native Rebecca panel API. This document replaces the former FastAPI /docs page and is intended for dashboard, bot, and third-party client developers.\n\nAuthentication: protected routes use `Authorization: Bearer <token>`. The token can be the JWT returned by `POST /api/admin/token`; admin API keys are also accepted by the Go auth layer where bearer validation is used. Public subscription and home/docs routes explicitly set `security: []`.\n\nResponse compatibility: most endpoints intentionally keep the JSON shape expected by the current dashboard. Some schemas are documented as generic JSON while the migration to strict OpenAPI component schemas continues.\n\nMutation behavior: user, service, node, config, and host mutations are designed to be transactional. When runtime sync is required, Rebecca writes `node_operations` in the same transaction so failed queue writes roll back the database change."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Rebecca panel origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/hosts": {
      "get": {
        "tags": [
          "Hosts"
        ],
        "summary": "List hosts",
        "description": "Return subscription hosts grouped by inbound tag.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an authenticated admin with host/service permissions. Host updates can affect service config output.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getHosts"
      },
      "put": {
        "tags": [
          "Hosts"
        ],
        "summary": "Update hosts",
        "description": "Bulk modify host definitions and service host links. Host order is now controlled by service host assignments, not by standalone host records.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an authenticated admin with host/service permissions. Host updates can affect service config output.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "putHosts"
      }
    },
...
    "/hosts": {
      "get": {
        "tags": [
          "Hosts"
        ],
        "summary": "List hosts alias",
        "description": "Compatibility alias for /api/hosts.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an authenticated admin with host/service permissions. Host updates can affect service config output.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getHosts"
      },
      "put": {
        "tags": [
          "Hosts"
        ],
        "summary": "Update hosts alias",
        "description": "Compatibility alias for /api/hosts.\n\nAuthentication: send `Authorization: Bearer <token>` using a JWT from `POST /api/admin/token` or a valid admin API key.\n\nPermissions: Requires an authenticated admin with host/service permissions. Host updates can affect service config output.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "putHosts"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or Rebecca admin API key"
      }
    },
...
    "requestBodies": {
      "JSON": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        },
        "description": "Generic JSON request body. See each operation description for required business fields and side effects."
      },
...
    },
    "responses": {
      "JSON": {
        "description": "JSON response. Exact object fields follow the dashboard-compatible Rebecca API response for the route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnyValue"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Object": {
        "type": "object",
        "additionalProperties": true
      },
      "AnyValue": {
        "nullable": true,
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "integer"
          },
          {
            "type": "boolean"
          }
        ]
      },
...
    }
  }
}